Update 8
Okay, finally figure out to make the rig compatible with the rig controls in UE5! It's not perfect of course but it is better than nothing so far!
Also I will have to delete some files within the project and folder soon. But that will be tomorrows job once I get the controls done!