Update 9
Finally finished the control rig. Trying to figure out how to animate this and blueprint it for the character. Other than that I have to delete some old assets that are no longer needed in the project and I have to import the character into the main project soon as well!