Small Update 10
Sorry for the small update, I ended busy with other projects today. But I managed to change the material to the coins and adding a shine and metallic texture to it. Also an important note I've turned off motion blur and change the Default RHI from DirectX 12 to DirectX 11.

Tomorrow I will work on the hud, add the character model to player and making animations too