Update 11
Finally exported and switch the mannequin with our main character and fully animated. Made the coins rotate in game and blueprint the hud so you can see how many the player has collected. And change the font design in the huds as well (still a work in process though!)

Hope this will push into git alright I've may messed up a bit too much when it  came to copying files, we'll see how it goes.