Hud Update
Change the main background for the main menu for our project. As well for updating images for the instruction menu and adding colour towards both the instruction and setting menus. Next I need to focus on the sky, pause button, and adding terrain for the second level. I would like to add enemies but I don't think we have time for that anymore